feat: move user and permissions components in api package
