+++ gravitee-apim-console-webui/src/management/api/user-group-access/groups/api-general-groups.component.spec.ts
@@ -24,10 +24,10 @@ import { ActivatedRoute } from '@angular/router';
 import { ApiGeneralGroupsComponent } from './api-general-groups.component';
 import { ApiGeneralGroupsHarness } from './api-general-groups.harness';
 
-import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../../../shared/testing';
-import { ApiGeneralUserGroupModule } from '../api-general-user-group.module';
-import { Api, fakeApiV1, fakeApiV2, fakeApiV4, fakeGroup, Group } from '../../../../../entities/management-api-v2';
-import { GioTestingPermissionProvider } from '../../../../../shared/components/gio-permission/gio-permission.service';
+import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../../shared/testing';
+import { ApiUserGroupModule } from '../api-user-group.module';
+import { Api, fakeApiV1, fakeApiV2, fakeApiV4, fakeGroup, Group } from '../../../../entities/management-api-v2';
+import { GioTestingPermissionProvider } from '../../../../shared/components/gio-permission/gio-permission.service';
 
 describe('ApiPortalGroupsComponent', () => {
   const API_ID = 'api-id';
@@ -38,7 +38,7 @@ describe('ApiPortalGroupsComponent', () => {
 
   const init = async (permissions: string[]) => {
     await TestBed.configureTestingModule({
-      imports: [NoopAnimationsModule, GioHttpTestingModule, ApiGeneralUserGroupModule, MatIconTestingModule],
+      imports: [NoopAnimationsModule, GioHttpTestingModule, ApiUserGroupModule, MatIconTestingModule],
       providers: [
         { provide: ActivatedRoute, useValue: { snapshot: { params: { apiId: API_ID } } } },
         { provide: GioTestingPermissionProvider, useValue: permissions },
