+++ gravitee-apim-console-webui/src/management/application/details/user-group-access/application-general-user-group.module.ts
@@ -44,7 +44,7 @@ import { GioPermissionModule } from '../../../../shared/components/gio-permissio
 import { GioUsersSelectorModule } from '../../../../shared/components/gio-users-selector/gio-users-selector.module';
 import { GioFormUserAutocompleteModule } from '../../../../shared/components/gio-user-autocomplete/gio-form-user-autocomplete.module';
 import { GioTableWrapperModule } from '../../../../shared/components/gio-table-wrapper/gio-table-wrapper.module';
-import { ApiGeneralUserGroupModule } from '../../../api/general/user-group-access/api-general-user-group.module';
+import { ApiUserGroupModule } from '../../../api/user-group-access/api-user-group.module';
 
 @NgModule({
   declarations: [ApplicationGeneralGroupsComponent, ApplicationGeneralMembersComponent, ApplicationGeneralTransferOwnershipComponent],
@@ -75,7 +75,7 @@ import { ApiGeneralUserGroupModule } from '../../../api/general/user-group-acces
     GioFormSlideToggleModule,
     GioUsersSelectorModule,
     GioTableWrapperModule,
-    ApiGeneralUserGroupModule,
+    ApiUserGroupModule,
     FormsModule,
   ],
 })
